+++ lib/tests/test_revset.rs
@@ -1420,6 +1420,64 @@ fn test_evaluate_expression_roots() {
 }
 
 #[test]
+fn test_evaluate_expression_forks() {
+    let test_repo = TestRepo::init();
+    let repo = &test_repo.repo;
+
+    let root_commit = repo.store().root_commit();
+    let mut tx = repo.start_transaction();
+    let mut_repo = tx.repo_mut();
+    /*
+     *   9       <- merge
+     *  /|\
+     * 6 7 8
+     *  \|/
+     *   5       <- 3-way fork
+     *  /|
+     * 3 4     <- not a fork
+     * |/
+     * 1 2     <- 2-way fork
+     * |/
+     * 0       <- 2-way fork from root
+     */
+    let commit1 = write_random_commit(mut_repo);
+    let commit2 = write_random_commit(mut_repo);
+    let commit3 = write_random_commit_with_parents(mut_repo, &[&commit1]);
+    let commit4 = write_random_commit_with_parents(mut_repo, &[&commit1]);
+    let commit5 = write_random_commit_with_parents(mut_repo, &[&commit3, &commit4]);
+    let commit6 = write_random_commit_with_parents(mut_repo, &[&commit5]);
+    let commit7 = write_random_commit_with_parents(mut_repo, &[&commit5]);
+    let commit8 = write_random_commit_with_parents(mut_repo, &[&commit5]);
+    let _commit9 = write_random_commit_with_parents(mut_repo, &[&commit6, &commit7, &commit8]);
+
+    // In the above graph, the forks are 0, 1, and 5
+    assert_eq!(
+        resolve_commit_ids(mut_repo, "forks()"),
+        vec![
+            commit5.id().clone(),
+            commit1.id().clone(),
+            root_commit.id().clone(),
+        ]
+    );
+
+    // 5 is a merge and a fork
+    assert_eq!(
+        resolve_commit_ids(mut_repo, "merges() & forks()"),
+        vec![commit5.id().clone()]
+    );
+
+    mut_repo.record_abandoned_commit(&commit2);
+    mut_repo.record_abandoned_commit(&commit7);
+    mut_repo.rebase_descendants().block_on().unwrap();
+
+    // After abandoning 2 and 7, only the root is no longer a fork.
+    assert_eq!(
+        resolve_commit_ids(mut_repo, "forks()"),
+        vec![commit5.id().clone(), commit1.id().clone(),]
+    );
+}
+
+#[test]
 fn test_evaluate_expression_parents() -> TestResult {
     let test_workspace = TestWorkspace::init();
     let repo = &test_workspace.repo;